Revert "move base AGN classes into DatabaseEmulator.py"
This reverts commit 0123516fd6faa5aa14b2a0aa8ef9dc73936a5a11.
